+++ doc/changelog.md
@@ -1,12 +1,15 @@
-Changes for 3.0.3
+Changes for 3.1.0
 =================
 
 - Updated to Gradle 8.6 and replaced the maven publisher with vanniktech's maven publisher to be
   compatible with Sonatype's new publishing model.
 
-- Made the plugin compatable with Gradle's
+- Made the plugin compatible with Gradle's
   [Configuration Cache](https://docs.gradle.org/current/userguide/configuration_cache.html) with
   thinks to @Nouran-11 and @abstratt.  Resolves #166 and #176
+
+- Made the plugin compatible with Gradle 9, with thanks to Alexander Zhukau (@zhukau).  Resolves
+  #173 and #177
 
 Changes for 3.0.2
 =================